email confirmation added

+++ Infrastructures/TravelManagmentSystem.Infrastructure/Services/CustomerService.cs
@@ -1,6 +1,8 @@
 ﻿using AutoMapper;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Storage.Json;
+using Microsoft.Extensions.Logging;
+using System.Text.Json;
 using TravelManagmentSystem.Application.Contracts.Repository;
 using TravelManagmentSystem.Application.Contracts.Services;
 using TravelManagmentSystem.Application.Contracts.UnitOfWorks;
@@ -15,28 +17,37 @@ namespace TravelManagmentSystem.Infrastructure.Services
         private readonly IGenericRepository<Customer> _customerRepository;
         private readonly IMapper _mapper;
         private readonly IUnitOfWork _unitOfWork;
-        public CustomerService(IGenericRepository<Customer> customerRepository, IMapper mapper, IUnitOfWork unitOfWork)
+        private readonly ILogger<CustomerService> _logger;
+        public CustomerService(IGenericRepository<Customer> customerRepository, IMapper mapper, IUnitOfWork unitOfWork, ILogger<CustomerService> logger)
         {
             _customerRepository = customerRepository;
             _mapper = mapper;
             _unitOfWork = unitOfWork;
+            _logger = logger;
         }
 
         public async Task<CustomerDto> AddAsync(CustomerDto dto)
         {
+            var dtoInputToJson=JsonSerializer.Serialize(dto);
+            _logger.LogInformation($"Customer Input: {dtoInputToJson}");
             var entityToCreate = _mapper.Map<Customer>(dto);
             await _customerRepository.AddAsync(entityToCreate);
             await _unitOfWork.SaveChangesAsync();
-            return _mapper.Map<CustomerDto>(entityToCreate);
+            var outDto= _mapper.Map<CustomerDto>(entityToCreate);
+            var dtoOutputToJson = JsonSerializer.Serialize(dto);
+            _logger.LogInformation($"Customer Output: {outDto}");
+            return outDto;
         }
 
         public async Task<IList<CustomerDto>> GetAllAsync()
         {
-            throw new Exception("Custom exception"); 
+            //throw new Exception("Custom exception"); 
 
             var query = _customerRepository.GetAll();
             var entities = await query.ToListAsync();
             var dtos = _mapper.Map<List<CustomerDto>>(entities);
+            var dtoOutputToJson = JsonSerializer.Serialize(dtos);
+            _logger.LogInformation($"Customer Get Output: {dtoOutputToJson}");
             return dtos;
         }
 
@@ -55,6 +66,10 @@ namespace TravelManagmentSystem.Infrastructure.Services
         public async Task<CustomerDto> RemoveAsync(Guid id )
         {
             var entity = await _customerRepository.GetAsync(id);
+            if (entity is null)
+            {
+                throw new NotFoundException($"This {id} is not found");
+            }
             _customerRepository.Remove(entity);
             _unitOfWork.SaveChanges();
             return _mapper.Map<CustomerDto>(entity);
@@ -63,7 +78,11 @@ namespace TravelManagmentSystem.Infrastructure.Services
         public async Task<CustomerDto> UpdateAsync(Guid id,CustomerDto dto)
         {
             var entity = await _customerRepository.GetAsync(id);
-      
+            if (entity is null)
+            {
+                throw new NotFoundException($"This {id} is not found");
+            }
+
             var entityToUpdate = _mapper.Map<Customer>(dto);
             entityToUpdate.Id = id;
             _customerRepository.Update(entityToUpdate);
